+++ CHANGES.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added --ecn server option to ship ECN bits (currently IPv6 only)
 - Move to Go 1.25, x/net 0.56 and x/sys 0.46
+- Allow IPv6 addresses without brackets
 
 ## 0.9.1 - 2021-05-18